+++ app/src/main/java/be/omnuzel/beatshare/db/UserDAO.java
@@ -145,13 +145,13 @@ public class UserDAO implements DataAccessObject<User> {
     }
 
     @Override
-    public void delete(int id) {
+    public void delete(long id) {
         db.delete(TABLE_NAME, COLUMN_ID + "=" + id, null);
         Log.i("USERDAO", "User with id : " + id + " has been deleted");
     }
 
     @Override
-    public User get(int id) {
+    public User get(long id) {
         Cursor c = db.query(TABLE_NAME, null, COLUMN_ID + "=" + id, null, null, null, null);
 
         if (c.getCount() > 0) {
